css/css-values/integer_interpolation_round_half_towards_positive_infinity_z_index.html

<!doctype html>
<title>Testing if integer interpolation is rounded towards positive infinity</title>
<link rel="author" title="Joonghun Park" href="pjh0718@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#combine-integers">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

#anim-target {
  animation: anim-z 4s steps(4) forwards 1;
  animation-delay: -1s;
  animation-play-state: paused;
}

@property --z {
  syntax: '<number>';
  initial-value: 0;
  inherits: true;
}

@keyframes anim-z {
  from {
    z-index: -2;
    --z: -2;
  }

  to {
    z-index: 0;
    --z: 0;
  }
}

</style>
<div id="anim-target"></div>
<script>
var test_description = "Integer interpolation should be rounded towards positive infinity";
test(
    t => {
        const target = document.getElementById("anim-target");
        const z_value = Number.parseFloat(getComputedStyle(target).getPropertyValue('--z'));
        const z_index_value = Number.parseFloat(getComputedStyle(target).getPropertyValue('z-index'));

        const rounded_z_value = Math.round(z_value);
        assert_equals(z_index_value, rounded_z_value, "Interpolation result for z-index should be rounded towards positive infinity");
    },
    test_description
);
</script>
